backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Password Hashing
bcrypt==4.1.2

# Validation & Serialization
marshmallow==3.20.1
python-dotenv==1.0.0

# PDF Processing
PyMuPDF==1.22.5
pdfplumber==0.10.3

# Machine Learning & Data Science
scikit-learn==1.4.2
pandas==2.1.4
numpy==1.26.2

# NLP (Optional)
regex==2023.10.3

# Task Queue (Optional)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development
autopep8==2.0.4
pylint==3.0.3